backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
psycopg2-binary
python-dotenv 
email-validator
python-multipart
bcrypt==4.0.1
passlib==1.7.4